% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdpq.R
\name{rdpq}
\alias{rdpq}
\title{Get r, d, p, and q functions from a given PDF and CDF.}
\usage{
rdpq(pdf, cdf)
}
\arguments{
\item{pdf}{the probability density function}

\item{cdf}{the cumulative distribution function}
}
\value{
r,d,p,q functions
}
\description{
This function returns the p, q, d, and r functions corresponding to the input PDF and CDF.
}
\examples{
\donttest{
rdpq(((1/gamma(delta))*((-log(1-((theta*(exp(-x^lambda))^v)/(1-(1-theta)*(exp(-x^lambda))^v))^(1/v)))^(delta-1))*((theta^(1/v))*lambda*(x^(lambda-1))*exp(-x^lambda))/((1-(1-theta)*(exp(-x^lambda))^v)^(1+(1/v)))),1-pgamma(-log(1-((theta*(exp(-x^lambda))^v)/(1-(1-theta)*(exp(-x^lambda))^v))^(1/v)),delta))
}


}
